listing bill modifies

+++ src/app/page/listing/listing-bill/listing-bill.component.ts
@@ -15,8 +15,6 @@ import { StatisticsService } from 'src/app/service/statistics.service';
 export class ListingBillComponent implements OnInit {
 
   @Input() id: number = 0;
-  // @Output() onUpdate: EventEmitter<Bill> = new EventEmitter();
-  // @Output() onDelete: EventEmitter<Bill> = new EventEmitter();
 
   billList$: BehaviorSubject<Bill[]> = this.billService.list$;
 
@@ -63,21 +61,11 @@ export class ListingBillComponent implements OnInit {
   }
 
   onRemove(bill: Bill): void {
-    // this.billService.remove(bill),
-    //   this.router.navigate(['/bills']);
-    // this.onDelete.emit(bill);
+    if (!confirm(`Biztosan törli ezt a számlát?
+    (id: ${bill.id} RendelésID: ${bill.orderID} Összeg: ${bill.amount})`)) {
+      return
+    }
 
-
-    // if (bill.id === 0) {
-    //   this.billService.create(bill).subscribe(
-    //     () => {
-    //       this.toastr.success('Sikeres a termék törlése!', 'Törlés!', { timeOut: 3000 });
-    //       this.router.navigate(['/bills']);
-    //     },
-    //     error => this.toastr.error('Hiba a termék törlésekor!', 'Hiba!', { timeOut: 3000 })
-    //   )
-    // }
-    // else {}
 
     this.billService.remove(bill).subscribe(
       () => {
